refactor: reorganize UI components into islands/ui/ directory (#47)
- Move Button, Checkbox, Switch from islands/ to islands/ui/
- Move TextInput from components/ to islands/ui/ for consistency
- Update all import paths to use @/islands/ui/ format
- Remove empty components/ directory

This establishes a clear structure for base UI components in islands/ui/
while keeping feature-specific islands in islands/.
