+++ app/islands/login-form.tsx
@@ -1,9 +1,9 @@
 import { useState } from "hono/jsx";
 import type { JSX } from "hono/jsx/jsx-runtime";
-import { TextInput } from "@/components/text-input";
+import Button from "@/islands/ui/button";
+import Checkbox from "@/islands/ui/checkbox";
+import { TextInput } from "@/islands/ui/text-input";
 import { apiClient } from "@/utils/api-client";
-import Button from "./button";
-import Checkbox from "./checkbox";
 
 export default function LoginForm(): JSX.Element {
 	const [email, setEmail] = useState("");
