+++ app/islands/signup-form.tsx
@@ -1,8 +1,8 @@
 import { useState } from "hono/jsx";
 import type { JSX } from "hono/jsx/jsx-runtime";
-import { TextInput } from "@/components/text-input";
+import Button from "@/islands/ui/button";
+import { TextInput } from "@/islands/ui/text-input";
 import { apiClient } from "@/utils/api-client";
-import Button from "./button";
 
 export default function SignupForm(): JSX.Element {
 	const [email, setEmail] = useState("");
